add retry to improve the reliability of build requests
